Finish abstracting calibration implementations to estimators submodule
