--- src/lf2i/calibration/parametric_cd.py
@@ -1,11 +1,0 @@
-# Thin re-export for backward compatibility.
-# Canonical implementations live in lf2i.estimators.torch_utils.cdf
-from lf2i.estimators.torch_utils.cdf import (   # noqa: F401
-    BrierScoreLoss,
-    WeightedBrierScoreLoss,
-    WeightedPinballLoss,
-    MomentRegressionLoss,
-    SigmoidCDF,
-    BetaNetwork,
-    ParametricCDFEstimator,
-)
