+++ src/lf2i/estimators/torch_utils/quantile_regressor.py
@@ -1,143 +1,201 @@
-from typing import Union, Dict, Any, Sequence
+from typing import Union, Dict, Any, Sequence, Optional, List
 import warnings
 
 import numpy as np
 import torch
+from torch.nn.functional import sigmoid
+from tqdm import tqdm
 from sklearn.model_selection import RandomizedSearchCV
 from sklearn.metrics import make_scorer, mean_pinball_loss
 from catboost import CatBoostRegressor
 
-from lf2i.calibration.torch_utils import QuantileLoss, FeedForwardNN, LearnerRegression
 from lf2i.utils.calibration_diagnostics_inputs import preprocess_train_quantile_regression
 from lf2i.utils.miscellanea import select_n_jobs
 from lf2i.estimators import AbstractQuantileRegressor
 
 
-def multi_quantile_mean_pinball_loss(
-    y_true: np.ndarray,
-    y_pred: np.ndarray,
-    alpha: Union[float, Sequence[float]]
-) -> float:
-    if isinstance(alpha, float):
-        return mean_pinball_loss(y_true=y_true, y_pred=y_pred, alpha=alpha)
-    else:
-        return np.mean([mean_pinball_loss(y_true=y_true, y_pred=y_pred[:, i], alpha=a) for i, a in enumerate(alpha)])
-
-
-def train_qr_algorithm(
-    test_statistics: Union[np.ndarray, torch.Tensor],
-    parameters: Union[np.ndarray, torch.Tensor],
-    algorithm: Union[str, Any],
-    alpha: Union[float, Sequence[float]],
-    param_dim: int,
-    algorithm_kwargs: Union[Dict[str, Any], Dict[str, Dict[str, Any]]] = {},
-    verbose: bool = True,
-    n_jobs: int = -2
-) -> Any:
-    """Dispatcher to train different quantile regressors and estimate critical values.
+class QuantileLoss(torch.nn.Module):
+    """Quantile loss as a PyTorch module.
+    Note that, although it supports multiple quantiles, there is currently no explicit constraint on their monotonicity to avoid quantile crossings.
 
     Parameters
     ----------
-    test_statistics : Union[np.ndarray, torch.Tensor]
-        The i-th element is the test statistics evaluated on the i-th element of `poi` (i.e., :math:`\theta_i`) and on :math:`x \sim F_{\theta_i}`.
-    parameters : Union[np.ndarray, torch.Tensor]
-        Parameters of interest in the calibration set.
-    algorithm : Union[str, Any]
-        Either 'cat-gb' for gradient boosted trees, 'nn' for a feed-forward neural network, or a custom algorithm (Any).
-        The latter must implement the `fit(X=..., y=...)` method.
-    alpha : Union[float, Sequence[float]]
-        The alpha quantile of the test statistic to be estimated. 
-        E.g., for 90% confidence intervals, it should be 0.9 if the acceptance region of the test statistic is on the left of the critical value. 
-        Similarly, it should be 0.1 if the acceptance region of the test statistic is on the right of the critical value. 
-        Must be in the range `(0, 1)`. NOTE: There is currently no explicit way of avoiding quantile crossings when estimating multiple quantiles simultaneously.
-    param_dim: int
-        Dimensionality of the parameter.
-    algorithm_kwargs : Union[Dict[str, Any], Dict[str, Dict[str, Any]]], optional
-        Keyword arguments for the desired algorithm, by default {}.
-        If algorithm == 'nn', then 'hidden_layer_shapes', 'epochs' and 'batch_size' must be present.
-        If algorithm == 'cat-gb', pass {'cv': hp_dist} to do a randomized search over the hyperparameters in hp_dist (a `Dict`) via 5-fold cross validation. 
-        Include 'n_iter' as a key to decide how many hyperparameter setting to sample for randomized search. Defaults to 10.
-    verbose: bool, optional
-        Whether to print information on the hyper-parameter search for quantile regression, by default True.
-    n_jobs : int, optional
-        Number of workers to use when doing random search with 5-fold CV. By default -2, which uses all cores minus one. If -1, use all cores.
-        `n_jobs == -1` uses all cores. If `n_jobs < -1`, then `n_jobs = os.cpu_count()+1+n_jobs`.
-
-    Returns
-    -------
-    Any
-        Fitted quantile regressor.
-
-    Raises
-    ------
-    ValueError
-        Only one of 'cat-gb', 'nn' or an instantiated custom quantile regressor (Any) is currently accepted as algorithm.
+    quantiles : Sequence[float]
+        Target quantiles. Values must be in the range `(0, 1)`.
     """
-    if isinstance(alpha, Sequence):
-        warnings.warn('You passed a sequence of levels alpha. Note that the current implementation cannot ensure quantile crossings will not happen.')
-    n_jobs = select_n_jobs(n_jobs)
+    def __init__(
+        self,
+        quantiles: Sequence[float]
+    ) -> None:
+        super().__init__()
+        self.quantiles = quantiles
 
-    if isinstance(algorithm, str):
-        if algorithm == "cat-gb":
-            if isinstance(alpha, Sequence):
-                alpha_string = str(alpha).replace('[', '').replace(']', '').replace(' ', '')
-            if 'cv' in algorithm_kwargs:
-                algorithm = RandomizedSearchCV(
-                    estimator=CatBoostRegressor(
-                        loss_function=f'Quantile:alpha={alpha}' if isinstance(alpha, float) else f'MultiQuantile:alpha={alpha_string}',
-                        silent=True,
-                        allow_writing_files=True if verbose else False
-                    ),
-                    param_distributions=algorithm_kwargs['cv'],
-                    n_iter=10 if 'n_iter' not in algorithm_kwargs else algorithm_kwargs['n_iter'],
-                    scoring=make_scorer(multi_quantile_mean_pinball_loss, alpha=alpha, greater_is_better=False),
-                    n_jobs=n_jobs,
-                    refit=True,
-                    cv=5,
-                    verbose=1 if verbose else 0
-                )
-            else:
-                algorithm = CatBoostRegressor(
-                    loss_function=f'Quantile:alpha={alpha}' if isinstance(alpha, float) else f'MultiQuantile:alpha={alpha_string}',
-                    **algorithm_kwargs
-                )
-            test_statistics, parameters = preprocess_train_quantile_regression(test_statistics, parameters, param_dim, algorithm)
-            algorithm.fit(X=parameters, y=test_statistics)
-        elif algorithm == 'nn':
-            # TODO: implement some form of hyperparameter tuning
-            quantiles = [alpha] if isinstance(alpha, float) else alpha
-            nn_kwargs = {arg: algorithm_kwargs[arg] for arg in ['hidden_activation', 'dropout_p', 'batch_norm'] if arg in algorithm_kwargs}
-            feedforward_nn = FeedForwardNN(
-                input_d=parameters.shape[1], 
-                output_d=len(quantiles),
-                hidden_layer_shapes=algorithm_kwargs.get('hidden_layer_shapes', [64, 64]),
-                **nn_kwargs
-            )
-            algorithm = LearnerRegression(
-                model=feedforward_nn, 
-                optimizer=torch.optim.Adam, 
-                loss=QuantileLoss(quantiles=quantiles), 
-                device="cuda" if torch.cuda.is_available() else "cpu",
-                verbose=verbose
-            )
-            test_statistics, parameters = preprocess_train_quantile_regression(test_statistics, parameters, param_dim, algorithm)
+    def forward(
+        self,
+        input: torch.Tensor,
+        target: torch.Tensor
+    ) -> torch.Tensor:
+        assert not target.requires_grad
+        assert input.size(0) == target.size(0)
+        losses = []
+        for i, q in enumerate(self.quantiles):
+            errors = target - input[:, i]
+            # "check" function
+            losses.append(torch.max((q - 1) * errors, q * errors).unsqueeze(1))
+        loss = torch.mean(torch.sum(torch.cat(losses, dim=1), dim=1))
+        return loss
 
-            learner_kwargs = {arg: algorithm_kwargs[arg] for arg in ['epochs', 'batch_size'] if arg in algorithm_kwargs}
-            if 'epochs' not in learner_kwargs:
-                learner_kwargs['epochs'] = 100
-            if 'batch_size' not in learner_kwargs:
-                learner_kwargs['batch_size'] = 64
 
-            algorithm.fit(X=parameters, y=test_statistics, **learner_kwargs)
-        else:
-            raise ValueError(f"Only 'cat-gb', 'nn' or custom algorithm (Any) are currently supported, got {algorithm}")
-    else:
-        if not isinstance(algorithm, AbstractQuantileRegressor):
-            raise ValueError(
-                f"Custom algorithm must implement AbstractQuantileRegressor "
-                f"(fit(X, y) and predict(X)), got {type(algorithm).__name__}."
-            )
-        test_statistics, parameters = preprocess_train_quantile_regression(test_statistics, parameters, param_dim, algorithm)
-        algorithm.fit(X=parameters, y=test_statistics)
+class FeedForwardNN(torch.nn.Module):
+    """Fully connected neural network.
 
-    return algorithm
+    Parameters
+    ----------
+    input_d : int
+        Dimensionality of the input.
+    output_d : int
+        Dimensionality of the output.
+    hidden_layer_shapes : Sequence[int]
+        The i-th element represents the number of neurons in the i-th hidden layer.
+    dropout_p : float, optional
+        Probability for the dropout layers, by default 0.0 (i.e., no dropout.)
+    batch_norm : bool, optional
+        Whether to apply batch normalization between each hidden layer or not.
+    """
+    def __init__(
+        self,
+        input_d: int,
+        output_d: int,
+        hidden_layer_shapes: Sequence[int],
+        hidden_activation: torch.nn.Module = torch.nn.ReLU(),
+        dropout_p: Optional[float] = None,
+        batch_norm: bool = False
+    ) -> None:
+        super().__init__()
+        self.input_d = input_d
+        self.output_d = output_d
+        self.hidden_layer_shapes = hidden_layer_shapes
+        self.hidden_activation = hidden_activation
+
+        self.build_model(batch_norm, dropout_p)
+
+    def build_model(self, batch_norm: bool, dropout_p: Optional[float]) -> None:
+        # input
+        self.model = [torch.nn.Linear(self.input_d, self.hidden_layer_shapes[0]), self.hidden_activation]
+        if batch_norm:
+            self.model += [torch.nn.BatchNorm1d(self.hidden_layer_shapes[0])]
+        if dropout_p:
+            self.model += [torch.nn.Dropout(p=dropout_p)]
+
+        # hidden
+        for i in range(0, len(self.hidden_layer_shapes)-1):
+            self.model += [torch.nn.Linear(self.hidden_layer_shapes[i], self.hidden_layer_shapes[i+1]), self.hidden_activation]
+            if batch_norm:
+                self.model += [torch.nn.BatchNorm1d(self.hidden_layer_shapes[i+1])]
+            if dropout_p:
+                self.model += [torch.nn.Dropout(p=dropout_p)]
+        # output: no sigmoid cause we use BCEWithLogitsLoss (more numerically stable thanks to log-sum-exp trick)
+        self.model += [torch.nn.Linear(self.hidden_layer_shapes[-1], self.output_d)]
+        self.model = torch.nn.Sequential(*self.model)
+
+    def forward(self, X: torch.Tensor) -> torch.Tensor:
+        return self.model(X)
+
+
+class Learner:
+    """Utility class to train a neural network.
+
+    Parameters
+    ----------
+    model : torch.nn.Module
+        Neural Network architecture.
+    optimizer : torch.optim.Optimizer
+        Chosen optimizer.
+    loss : torch.nn.Module
+        Loss function to minimize via SGD.
+    device : str, optional
+        Device on which to perform computations, by default "cpu"
+    """
+    def __init__(
+        self,
+        model: torch.nn.Module,
+        optimizer: torch.optim.Optimizer,
+        loss: torch.nn.Module,
+        device: str = "cpu",
+        verbose: bool = True
+    ) -> None:
+        self.device = torch.device(device)
+        self.model = model.to(self.device)
+        self.optimizer = optimizer(self.model.parameters())
+        self.loss = loss.to(self.device)
+        self.loss_trajectory: List[np.ndarray] = []
+        self.verbose = verbose
+
+    def __getstate__(self):
+        # The optimizer is only needed during training; drop it to avoid pickling
+        # torch.backends.* ConfigModuleInstance objects that live in its closure chain.
+        state = self.__dict__.copy()
+        state['optimizer'] = None
+        return state
+
+    def __setstate__(self, state):
+        self.__dict__.update(state)
+
+    def fit(
+        self,
+        X: torch.Tensor,
+        y: torch.Tensor,
+        epochs: int,
+        batch_size: int
+    ) -> None:
+        self.model.train()
+        if self.verbose:
+            pbar = tqdm(total=epochs, desc="Training Neural Network")
+        for _ in range(epochs):
+            shuffled_idx = torch.randperm(X.shape[0])
+            X = X[shuffled_idx, :]
+            y = y[shuffled_idx]
+            epoch_losses = []
+            for idx in range(0, X.shape[0], batch_size):
+                self.optimizer.zero_grad()
+
+                batch_X = X[idx: min(idx + batch_size, X.shape[0]), :].float().to(self.device)
+                batch_y = y[idx: min(idx + batch_size, y.shape[0])].reshape(-1, 1).float().to(self.device)
+
+                batch_predictions = self.model(batch_X)
+                batch_loss = self.loss(input=batch_predictions, target=batch_y)
+                batch_loss.backward()
+                self.optimizer.step()
+                epoch_losses.append(batch_loss.cpu().detach().numpy())
+            if self.verbose:
+                pbar.update(1)
+            self.loss_trajectory.append(np.mean(epoch_losses))
+        if self.verbose:
+            pbar.close()
+
+    def predict(
+        self,
+        X: torch.Tensor
+    ) -> torch.Tensor:
+        raise NotImplementedError('Use the children classes that implement the correct methods for inference in regression or classification settings.')
+
+
+class LearnerRegression(Learner):
+
+    def predict(
+        self,
+        X: torch.Tensor
+    ) -> torch.Tensor:
+        self.model.eval()
+        return self.model(X.to(self.device)).cpu().detach()
+
+
+class LearnerClassification(Learner):
+
+    def predict_proba(
+        self,
+        X: torch.Tensor
+    ) -> torch.Tensor:
+        self.model.eval()
+        # output predicted probability for positive class
+        return sigmoid(self.model(X.float().to(self.device)).cpu().detach().reshape(X.shape[0], -1))
